Add b1gMail 7.5 core security, Tabler UI, routing, and webdisk updates
Introduce cookie-based sessions with CSRF protection, idle lock, and optional MFA (TOTP/email/backup codes). Migrate password hashing from MD5 to bcrypt/Argon2id on login and add configurable hash policy.

Replace sid-in-URL handling with pretty URL routing via front controllers and a plugin route registry. Ship a Tabler theme for LI/NLI and align the ACP with the same UI, plus PWA install support and user/admin avatars.

Extend webdisk with previews, gallery, in-browser editing, media playback, improved uploads/shares, and vcf/ics import. Update setup for 7.5, PHP 8.1+, and plugin APIs (CSRF, routing, web push).

+++ src/serverlib/filters.inc.php
@@ -316,8 +316,8 @@ class BMMailFilter_Bayes_Word
 		$allSpam 		= max($bayesSpam, 1);
 		$good 			= $this->inNonSpam * 2;
 		$bad 			= $this->inSpam;
-		$pGood 			= (double)min((double)$good / (double)$allNonSpam, (double)1);
-		$pBad 			= (double)min((double)$bad / (double)$allSpam, (double)1);
+		$pGood 			= (float) min((float) $good / (float) $allNonSpam, 1.0);
+		$pBad 			= (float) min((float) $bad / (float) $allSpam, 1.0);
 		$this->prob 	= max(min($pBad / ($pGood + $pBad), 0.9999), 0.0001);
 
 		// calculate interest
